시험분석 기능 고도화, 시스템 종료 조건 추가, 쿼리문 고도화, 응답 기준 변경

+++ src/main/java/com/chatbot/web/domains/User.java
@@ -8,12 +8,12 @@ import org.springframework.stereotype.Component;
 import java.time.LocalDate;
 import java.time.LocalDateTime;
 
-@Component @Lazy
-@Data
+@Component
+@Lazy @Data
 @Alias("user")
 @MappedTypes(LocalDate.class)
 public class User {
-    private Long id;
+    private int id;
     private int userCode, schoolCode, curGrade, homeClass, positionChecker;
     private String name, userId, userPw, email, phone, gardianPhone, gardianRelation, gender;
     private LocalDateTime insertDate, updateDate;
